requirements.txt
Flask
requests
fpdf
Pillow
google-cloud-aiplatform
google-cloud-vision
google-genai
weasyprint
reportlab